Add files via upload
Executing this Python file enables us to collect data on vehicle perception, including images, as well as vehicle commands such as throttle, steering, and braking. You can utilize either keyboard commands or Carla's built-in autopilot (add "--autopilot" to your command) within the Carla Simulator. The simulation provides visual representation of the vehicle's movement, and the gathered data is subsequently stored in two separate .npy files: "inputs" for the vehicle's perception and "outputs" for the vehicle's commands.